# backend/audit-service/docs/swagger.yaml
definitions:
  models.ActivityPlanResponse:
    properties:
      annual_plan_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      objective:
        type: string
      planned_end:
        type: string
      planned_start:
        type: string
      project_code:
        type: string
      scope:
        type: string
      status:
        type: string
      target_unit_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.AuditCharterResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      filename:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
...
  models.CreateActivityPlanRequest:
    properties:
      annual_plan_id:
        type: string
      objective:
        type: string
      planned_end:
        type: string
      planned_start:
        type: string
      project_code:
        maxLength: 50
        type: string
      scope:
        type: string
      status:
        maxLength: 50
        type: string
      target_unit_id:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - annual_plan_id
    - planned_end
    - planned_start
    - project_code
    - target_unit_id
    - title
    type: object
...
  models.CreateAuditCharterRequest:
    properties:
      content:
        type: string
      filename:
        maxLength: 200
        type: string
      title:
        maxLength: 200
        type: string
      version:
        maxLength: 20
        type: string
    required:
    - content
    - filename
    - title
    - version
    type: object
...
  models.UpdateActivityPlanRequest:
    properties:
      objective:
        type: string
      planned_end:
        type: string
      planned_start:
        type: string
      scope:
        type: string
      status:
        maxLength: 50
        type: string
      target_unit_id:
        type: string
      title:
        maxLength: 255
        type: string
    type: object
...
  models.UpdateAuditCharterRequest:
    properties:
      content:
        type: string
      filename:
        maxLength: 200
        type: string
      title:
        maxLength: 200
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorDetail'
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
  /api/v1/audit-activities:
    get:
      description: Get a list of audit activities with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: Annual Plan ID filter
        in: query
        name: annual_plan_id
        type: string
      - description: Target Unit ID filter
        in: query
        name: target_unit_id
        type: string
      - description: Status filter
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Audit Activities
      tags:
      - audit-activities
    post:
      consumes:
      - application/json
      description: Create a new audit activity
      parameters:
      - description: Create audit activity request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateActivityPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create Audit Activity
      tags:
      - audit-activities
  /api/v1/audit-activities/{id}:
    delete:
      description: Delete an audit activity
      parameters:
      - description: Audit Activity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Audit Activity
      tags:
      - audit-activities
    get:
      description: Get an audit activity by ID
      parameters:
      - description: Audit Activity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ActivityPlanResponse'
              type: object
      security:
      - Bearer: []
      summary: Get Audit Activity
      tags:
      - audit-activities
    put:
      consumes:
      - application/json
      description: Update an audit activity
      parameters:
      - description: Audit Activity ID
        in: path
        name: id
        required: true
        type: string
      - description: Update audit activity request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateActivityPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Audit Activity
      tags:
      - audit-activities
  /api/v1/audit-activities/project:
    get:
      description: Get an audit activity by project code
      parameters:
      - description: Project Code
        in: query
        name: project_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ActivityPlanResponse'
              type: object
      security:
      - Bearer: []
      summary: Get Audit Activity By Project Code
      tags:
      - audit-activities
...
  /api/v1/audit-charters:
    get:
      description: Get a list of audit charters with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      - description: Active status filter
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Audit Charters
      tags:
      - audit-charters
    post:
      consumes:
      - application/json
      description: Create a new audit charter
      parameters:
      - description: Create audit charter request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAuditCharterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create Audit Charter
      tags:
      - audit-charters
  /api/v1/audit-charters/{id}:
    delete:
      description: Delete an audit charter
      parameters:
      - description: Audit Charter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete Audit Charter
      tags:
      - audit-charters
    get:
      description: Get an audit charter by ID
      parameters:
      - description: Audit Charter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuditCharterResponse'
              type: object
      security:
      - Bearer: []
      summary: Get Audit Charter
      tags:
      - audit-charters
    put:
      consumes:
      - application/json
      description: Update an audit charter
      parameters:
      - description: Audit Charter ID
        in: path
        name: id
        required: true
        type: string
      - description: Update audit charter request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAuditCharterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update Audit Charter
      tags:
      - audit-charters
  /api/v1/audit-charters/{id}/activate:
    post:
      description: Set an audit charter as the active charter
      parameters:
      - description: Audit Charter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Set Active Audit Charter
      tags:
      - audit-charters
  /api/v1/audit-charters/active:
    get:
      description: Get the active audit charter
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuditCharterResponse'
              type: object
      security:
      - Bearer: []
      summary: Get Active Audit Charter
      tags:
      - audit-charters
...
swagger: "2.0"